requirements.txt
requests
discord
pyyaml
python-dotenv